chore: clarify default config endpoints for RPCs (#20)
## Summary

While reading the code I noticed we use `Default` a lot for the RPC
configuration, which is very confusing as it defaults to solana's public
devnet even in case of ephemeral related code.

## Details

This PR makes which `remote` we use inside of RPC configuration objects
to be forced be explicit instead of implicit. This allows surfacing
configuration errors hidden in the current codebase.